Changed Lambda-Notes
Originally had my front-end a submodule of git when I got an error trying to copy/paste it.  Then I realized I could just copy the app itself, not the whole folder.

This will allow me to modify it, whereas before it wouldn't (I assume) because it was just piggybacking off the repo.