JWT Working...sort of.
My problem that I'm having is that the token allows access to every user notes, given that the user id passed in as a parameter.

Maybe I shouldn't worry about this too much for the purpose of our project... but it is bugging me because I think it shows the underlying problem of me not understanding what I am doing...

controllers/users.js --- Login route works to login someone, at least in the sense it provides a token if the username/password match.  Logout is broken, and I have no clue how to fix it since it doesn't access the token if I don't pass it in the header, which I assume you wouldn't do.

middleware/loggedIn.js --- I had tried using sessions, and failed even more so than with tokens, so its back to the original.

routes/index.js --- put loggedIn middleware back in.

server.js --- Was trying to do sessions, had to put it back to JWT.