Still a mess. Working on making notes specific to users.
Front end portion is still a mess.
Broke some things that were working before.
Still need to reorganize a few things but currently focused on adding
to the backend.
Everything seems to work fine in postman. Notes now include username so
that we can retrieve only that specific user's notes.