Finally Have LogIn/LogOut Working, I hope...
I could be wrong, but the only way I can think to test whether it is working makes it seem like it is...

controllers/notes.js --- changed conditionals to be based on userId instead of username... I had a good(maybe)  reason for this, but the reason is eluding me at the moment...

controllers/users.js --- On logout, the token the loggedIn user was using is added to a seperate database.

middleware/loggedIn.js --- Check to see if the loggedOut token exists in BadToken database, if it does, does not allow user to do anything until logged in again.

models/badToken.js --- created schema/model for badToken database documents.

routes/index.js --- Re-implemented loggedIn for the millionth time, removed loggedOut middleware I was thinking of using.